refactor(grade): gradeCalculator を責務別6ファイルへ分割
1066行の gradeCalculator.ts を責務ごとに分割し、calculateGrades の
オーケストレーションのみ本体に残す（537行）。型規約に従い共有型は
独立ファイルへ集約し、単一ファイル内の型はファイル内宣言とした。

- gradeCalculatorTypes.ts: ExamDataCache / DataSourceInfo（共有型）
- rawScoreCalculator.ts:   getRawScore + coursework 系
- examScoreCalculator.ts:  exam_total / crop_region スコア算出
- absentEstimation.ts:     平均・回帰・OLS 推定 + applyAdjustmentAndClamp
- gradeLabel.ts:           determineGradeLabel

re-export 禁止方針に従い、移設した estimateAbsentScore /
applyAdjustmentAndClamp はテストの import 元を absentEstimation へ
直接更新。振る舞いは不変（typecheck + 30テスト + lint 全通過）。
