+++ electron-src/lib/shared/calculations/gradeCalculator.ts
@@ -17,27 +17,14 @@ import {
 } from "../../../../src/types/grade.types"
 import prisma from "../../prisma/client"
 import { computeLiveMaxScore } from "../../prisma/gradeDataSource"
-import { calculateActualScore } from "../../prisma/questionScore"
-import { resolveEffectiveScores } from "./scoreResolution"
 import {
-  calculateSubtotalScoreBySubtotalId,
-  type QuestionScoreForSubtotal,
-} from "./subtotalCalculator"
-
-interface ExamDataCache {
-  questionScores: QuestionScoreForSubtotal[]
-  cropRegions: { id: string; type: string; points: number | null }[]
-}
-
-interface DataSourceInfo {
-  id: string
-  maxScore: number
-  absentMethod: AbsentMethod
-  absentRatio: number
-  absentOffset: number
-  estimationMode: string
-  estimationSourceIds: string[]
-}
+  applyAdjustmentAndClamp,
+  estimateAbsentScore,
+} from "./absentEstimation"
+import type { DataSourceInfo, ExamDataCache } from "./gradeCalculatorTypes"
+import { determineGradeLabel } from "./gradeLabel"
+import { getRawScore } from "./rawScoreCalculator"
+import { resolveEffectiveScores } from "./scoreResolution"
 
 /**
  * 成績を算出する
@@ -547,520 +534,4 @@ export async function calculateGrades(gradeId: string): Promise<{
       error: error instanceof Error ? error.message : "Unknown error",
     }
   }
-}
-
-/**
- * DataSourceからrawScoreを取得（推定前の実スコア）
- */
-async function getRawScore(
-  studentId: string,
-  dataSource: {
-    type: string
-    examId: string | null
-    subtotalId: string | null
-    cropRegionId: string | null
-    courseworkItem?: {
-      maxScore: unknown
-      inputMode: string
-      scores: {
-        studentId: string
-        score: unknown
-        letterValue?: string | null
-        adjustment?: unknown
-      }[]
-      letterScales: { label: string; score: unknown }[]
-    } | null
-    coursework?: {
-      items: {
-        maxScore: unknown
-        inputMode: string
-        scores: {
-          studentId: string
-          score: unknown
-          letterValue?: string | null
-          adjustment?: unknown
-        }[]
-        letterScales: { label: string; score: unknown }[]
-      }[]
-    } | null
-  },
-  examDataCache: Map<string, ExamDataCache>
-): Promise<number | null> {
-  if (dataSource.type === "exam_total" && dataSource.examId) {
-    return calculateExamTotalScore(studentId, dataSource.examId, examDataCache)
-  } else if (
-    dataSource.type === "subtotal" &&
-    dataSource.subtotalId &&
-    dataSource.examId
-  ) {
-    const examData = examDataCache.get(dataSource.examId)
-    if (examData) {
-      const result = await calculateSubtotalScoreBySubtotalId(
-        studentId,
-        dataSource.subtotalId,
-        examData.questionScores,
-        examData.cropRegions as Parameters<
-          typeof calculateSubtotalScoreBySubtotalId
-        >[3]
-      )
-      return result.score
-    }
-  } else if (
-    dataSource.type === "crop_region" &&
-    dataSource.cropRegionId &&
-    dataSource.examId
-  ) {
-    return calculateCropRegionScore(
-      studentId,
-      dataSource.cropRegionId,
-      dataSource.examId,
-      examDataCache
-    )
-  } else if (dataSource.type === "coursework" && dataSource.courseworkItem) {
-    return getCourseworkRawScore(studentId, dataSource.courseworkItem)
-  } else if (dataSource.type === "coursework_total" && dataSource.coursework) {
-    return getCourseworkTotalRawScore(studentId, dataSource.coursework.items)
-  }
-  return null
-}
-
-/**
- * coursework_total型データソース（試験外成績資料の全評価項目合計）の実スコアを算出する。
- * 各評価項目のスコアを getCourseworkRawScore で求めて合算する。
- * exam_total と同様、採点済み項目のみを合算し、全項目が未入力なら null を返す。
- */
-function getCourseworkTotalRawScore(
-  studentId: string,
-  items: {
-    maxScore: unknown
-    inputMode: string
-    scores: {
-      studentId: string
-      score: unknown
-      letterValue?: string | null
-      adjustment?: unknown
-    }[]
-    letterScales: { label: string; score: unknown }[]
-  }[]
-): number | null {
-  let total = 0
-  let hasScored = false
-  for (const item of items) {
-    const score = getCourseworkRawScore(studentId, item)
-    if (score !== null) {
-      hasScored = true
-      total += score
-    }
-  }
-  return hasScored ? total : null
-}
-
-/**
- * coursework型データソース（試験外成績資料の評価項目）の実スコアを算出する。
- * - letterモード: 入力された評価記号を変換表で点数化
- * - numericモード: 入力された数値をそのまま使用
- * いずれも加点・減点(adjustment)を加算し、結果はクランプしない。
- * 実際に入力された得点は配点超え（100%超）も負値（減点で0未満）もそのまま反映する。
- * （推定で算出した代替スコアは別途 applyAdjustmentAndClamp で [0, 満点] に収める。
- * 実入力と推定値で扱いを分け、推定値だけが配点を超えないようにするため。）
- * 満点は評価項目（CourseworkItem.maxScore）を live 参照する。
- */
-function getCourseworkRawScore(
-  studentId: string,
-  item: {
-    maxScore: unknown
-    inputMode: string
-    scores: {
-      studentId: string
-      score: unknown
-      letterValue?: string | null
-      adjustment?: unknown
-    }[]
-    letterScales: { label: string; score: unknown }[]
-  }
-): number | null {
-  const courseworkScore = item.scores.find(
-    (score) => score.studentId === studentId
-  )
-  if (!courseworkScore) return null
-
-  // 基準スコア（変換前・加減点前）を決定
-  let base: number | null
-  if (item.inputMode === "letter") {
-    if (
-      courseworkScore.letterValue === null ||
-      courseworkScore.letterValue === undefined
-    ) {
-      base = null
-    } else {
-      const scale = item.letterScales.find(
-        (letterScale) => letterScale.label === courseworkScore.letterValue
-      )
-      base = scale ? Number(scale.score) : null
-    }
-  } else {
-    base =
-      courseworkScore.score !== null && courseworkScore.score !== undefined
-        ? Number(courseworkScore.score)
-        : null
-  }
-
-  if (base === null || Number.isNaN(base)) return null
-
-  const adjustment =
-    courseworkScore.adjustment !== null &&
-    courseworkScore.adjustment !== undefined
-      ? Number(courseworkScore.adjustment)
-      : 0
-  // クランプしない。配点超え・負値のいずれも入力どおり成績算出に反映する。
-  return base + adjustment
-}
-
-/**
- * 欠測時の代替スコアを推定
- * rawScoreMapには推定前の実スコアのみが格納されている（循環推定の防止）
- */
-export function estimateAbsentScore(
-  method: AbsentMethod,
-  studentId: string,
-  dataSourceId: string,
-  maxScore: number,
-  rawScoreMap: Map<string, Map<string, number | null>>,
-  allDataSources: DataSourceInfo[]
-): number | null {
-  if (method === "zero") {
-    return 0
-  }
-  if (method === "average") {
-    return estimateByAverage(
-      studentId,
-      dataSourceId,
-      maxScore,
-      rawScoreMap,
-      allDataSources
-    )
-  }
-  if (method === "regression") {
-    return estimateByRegression(
-      studentId,
-      dataSourceId,
-      maxScore,
-      rawScoreMap,
-      allDataSources
-    )
-  }
-  return null
-}
-
-/**
- * 平均比率推定: 同じ生徒の他DataSourceのスコア比率(score/maxScore)を平均
- * → 平均比率 × 当該DataSource.maxScore
- */
-function estimateByAverage(
-  studentId: string,
-  dataSourceId: string,
-  maxScore: number,
-  rawScoreMap: Map<string, Map<string, number | null>>,
-  allDataSources: DataSourceInfo[]
-): number | null {
-  const studentScores = rawScoreMap.get(studentId)
-  if (!studentScores) return null
-
-  let ratioSum = 0
-  let ratioCount = 0
-
-  for (const dataSource of allDataSources) {
-    if (dataSource.id === dataSourceId) continue
-    if (dataSource.maxScore <= 0) continue
-    const score = studentScores.get(dataSource.id)
-    if (score === null || score === undefined) continue
-    ratioSum += score / dataSource.maxScore
-    ratioCount++
-  }
-
-  if (ratioCount === 0) return null
-  const avgRatio = ratioSum / ratioCount
-  return clamp(avgRatio * maxScore, 0, maxScore)
-}
-
-/**
- * OLS重回帰法推定:
- * 他の生徒のデータを訓練データとして、他DataSourceスコアから
- * 当該DataSourceスコアを予測する重回帰モデルを構築。
- * β = (X^T X)^(-1) X^T Y で係数を算出し、対象生徒のスコアを予測。
- */
-function estimateByRegression(
-  studentId: string,
-  dataSourceId: string,
-  maxScore: number,
-  rawScoreMap: Map<string, Map<string, number | null>>,
-  allDataSources: DataSourceInfo[]
-): number | null {
-  // 他DataSourceのID一覧（predictor変数）
-  const predictorDsIds = allDataSources
-    .filter(
-      (dataSource) => dataSource.id !== dataSourceId && dataSource.maxScore > 0
-    )
-    .map((dataSource) => dataSource.id)
-
-  if (predictorDsIds.length === 0) return null
-
-  // 対象生徒のpredictor値を取得
-  const targetStudentScores = rawScoreMap.get(studentId)
-  if (!targetStudentScores) return null
-
-  // 対象生徒が持っているpredictorのみを使用
-  const availablePredictors = predictorDsIds.filter((id) => {
-    const score = targetStudentScores.get(id)
-    return score !== null && score !== undefined
-  })
-
-  if (availablePredictors.length === 0) return null
-
-  // 訓練データ収集: 他の生徒で、当該DSと全available predictorのスコアが揃っている行
-  const X: number[][] = [] // 各行 = [1, x1, x2, ...] (切片含む)
-  const Y: number[] = []
-
-  for (const [otherStudentId, scores] of rawScoreMap) {
-    if (otherStudentId === studentId) continue
-    const y = scores.get(dataSourceId)
-    if (y === null || y === undefined) continue
-
-    const row: number[] = [1] // 切片項
-    let complete = true
-    for (const predId of availablePredictors) {
-      const x = scores.get(predId)
-      if (x === null || x === undefined) {
-        complete = false
-        break
-      }
-      row.push(x)
-    }
-    if (!complete) continue
-
-    X.push(row)
-    Y.push(y)
-  }
-
-  // 最低でも説明変数+1のサンプルが必要（＋余裕を持って）
-  const minSamples = availablePredictors.length + 2
-  if (X.length < minSamples) {
-    // サンプル不足の場合、平均比率法にフォールバック
-    return estimateByAverage(
-      studentId,
-      dataSourceId,
-      maxScore,
-      rawScoreMap,
-      allDataSources
-    )
-  }
-
-  // OLS: β = (X^T X)^(-1) X^T Y
-  const p = X[0].length // パラメータ数（切片含む）
-  const beta = solveOLS(X, Y, p)
-  if (!beta) {
-    // 特異行列の場合、平均比率法にフォールバック
-    return estimateByAverage(
-      studentId,
-      dataSourceId,
-      maxScore,
-      rawScoreMap,
-      allDataSources
-    )
-  }
-
-  // 対象生徒のpredictor値で予測
-  const xTarget = [1]
-  for (const predId of availablePredictors) {
-    xTarget.push(targetStudentScores.get(predId)!)
-  }
-
-  let predicted = 0
-  for (let j = 0; j < p; j++) {
-    predicted += beta[j] * xTarget[j]
-  }
-
-  return clamp(predicted, 0, maxScore)
-}
-
-/**
- * OLS正規方程式を解く: β = (X^T X)^(-1) X^T Y
- * ガウス消去法で (X^T X) β = X^T Y を解く
- */
-function solveOLS(X: number[][], Y: number[], p: number): number[] | null {
-  const n = X.length
-
-  // X^T X (p×p)
-  const XtX: number[][] = Array.from({ length: p }, () => Array(p).fill(0))
-  for (let i = 0; i < p; i++) {
-    for (let j = 0; j < p; j++) {
-      let sum = 0
-      for (let k = 0; k < n; k++) {
-        sum += X[k][i] * X[k][j]
-      }
-      XtX[i][j] = sum
-    }
-  }
-
-  // X^T Y (p)
-  const XtY: number[] = Array(p).fill(0)
-  for (let i = 0; i < p; i++) {
-    let sum = 0
-    for (let k = 0; k < n; k++) {
-      sum += X[k][i] * Y[k]
-    }
-    XtY[i] = sum
-  }
-
-  // 拡大係数行列 [XtX | XtY] → ガウス消去法
-  const aug: number[][] = XtX.map((row, i) => [...row, XtY[i]])
-
-  for (let col = 0; col < p; col++) {
-    // ピボット選択
-    let maxRow = col
-    let maxVal = Math.abs(aug[col][col])
-    for (let row = col + 1; row < p; row++) {
-      if (Math.abs(aug[row][col]) > maxVal) {
-        maxVal = Math.abs(aug[row][col])
-        maxRow = row
-      }
-    }
-    if (maxVal < 1e-12) return null // 特異行列
-
-    // 行交換
-    if (maxRow !== col) {
-      ;[aug[col], aug[maxRow]] = [aug[maxRow], aug[col]]
-    }
-
-    // 前進消去
-    const pivot = aug[col][col]
-    for (let row = col + 1; row < p; row++) {
-      const factor = aug[row][col] / pivot
-      for (let j = col; j <= p; j++) {
-        aug[row][j] -= factor * aug[col][j]
-      }
-    }
-  }
-
-  // 後退代入
-  const beta = Array(p).fill(0)
-  for (let i = p - 1; i >= 0; i--) {
-    if (Math.abs(aug[i][i]) < 1e-12) return null
-    let sum = aug[i][p]
-    for (let j = i + 1; j < p; j++) {
-      sum -= aug[i][j] * beta[j]
-    }
-    beta[i] = sum / aug[i][i]
-  }
-
-  return beta
-}
-
-/**
- * 調整(ratio/offset)を適用し、[0, maxScore]にクランプ
- */
-export function applyAdjustmentAndClamp(
-  estimated: number,
-  ratio: number,
-  offset: number,
-  maxScore: number
-): number {
-  const adjusted = estimated * ratio + offset
-  return Math.round(clamp(adjusted, 0, maxScore) * 100) / 100
-}
-
-/**
- * 値を[min, max]範囲にクランプ
- */
-function clamp(value: number, min: number, max: number): number {
-  return Math.max(min, Math.min(max, value))
-}
-
-/**
- * exam_total: 試験の全QUESTION_ANSWER CropRegionスコア合計
- */
-function calculateExamTotalScore(
-  studentId: string,
-  examId: string,
-  examDataCache: Map<string, ExamDataCache>
-): number | null {
-  const examData = examDataCache.get(examId)
-  if (!examData) return null
-
-  const studentScores = examData.questionScores.filter(
-    (questionScore) => questionScore.studentId === studentId
-  )
-  const questionRegions = examData.cropRegions.filter(
-    (cropRegion) => cropRegion.type === "QUESTION_ANSWER"
-  )
-
-  let totalScore = 0
-  let hasScored = false
-
-  for (const cropRegion of questionRegions) {
-    const scoreData = studentScores.find(
-      (questionScore) => questionScore.cropRegionId === cropRegion.id
-    )
-    if (scoreData) {
-      const actualScore = calculateActualScore(
-        scoreData,
-        cropRegion.points ?? 0
-      )
-      if (actualScore !== null) {
-        hasScored = true
-        totalScore += actualScore
-      }
-    }
-  }
-
-  return hasScored ? totalScore : null
-}
-
-/**
- * crop_region: 単一CropRegionのスコア取得
- */
-function calculateCropRegionScore(
-  studentId: string,
-  cropRegionId: string,
-  examId: string,
-  examDataCache: Map<string, ExamDataCache>
-): number | null {
-  const examData = examDataCache.get(examId)
-  if (!examData) return null
-
-  const cropRegion = examData.cropRegions.find(
-    (cropRegion) => cropRegion.id === cropRegionId
-  )
-  if (!cropRegion) return null
-
-  const scoreData = examData.questionScores.find(
-    (questionScore) =>
-      questionScore.studentId === studentId &&
-      questionScore.cropRegionId === cropRegionId
-  )
-  if (!scoreData) return null
-
-  return calculateActualScore(scoreData, cropRegion.points ?? 0)
-}
-
-/**
- * パーセンテージから成績ラベルを決定（降順マッチ）
- */
-function determineGradeLabel(
-  percentage: number | null,
-  boundaries: { label: string; minPercentage: unknown; order: number }[]
-): string | null {
-  if (percentage === null || boundaries.length === 0) return null
-
-  const sorted = [...boundaries].sort(
-    (boundaryA, boundaryB) =>
-      Number(boundaryB.minPercentage) - Number(boundaryA.minPercentage)
-  )
-
-  for (const boundary of sorted) {
-    if (percentage >= Number(boundary.minPercentage)) {
-      return boundary.label
-    }
-  }
-  return sorted[sorted.length - 1]?.label ?? null
 }
